♻️ Refactor ProgressBanner out of Notifications (#849) * Move all progress bar-related code into its own component * Add the checkmark and cross icons to the progress banner
